server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "browser-use==0.9.7",
    "fastapi[standard]>=0.121.3",
    "openai-agents>=0.6.1",
    "playwright>=1.56.0",
    "pymongo>=4.6.0",
    "python-dotenv>=1.0.0",
    "requests==2.32.5",
    "uvicorn>=0.38.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
    "bcrypt==4.0.1",
]
